Revert "Implement new tile loading (1cdb0a3)"
Újra mûködik a betöltés-mentés, visszatértünk a régi kerékvágásba. Illetve annyit változtattam hogy kiszedtem a Utils-ból a GetResourcesNames metódust ami a tile textúrák path-ját adja vissza, úgysem használjuk máshol. Plusz a TileParser sem static már és így nem függ a Utils-tól sem.
