+++ EvoCarcassonne/Backend/Extensions.cs
@@ -17,23 +17,5 @@ namespace EvoCarcassonne.Backend
                 return value;
             }
         }
-
-        public static BitmapImage ToBitmapImage(this Bitmap bitmap)
-        {
-            using (var memory = new MemoryStream())
-            {
-                bitmap.Save(memory, ImageFormat.Png);
-                memory.Position = 0;
-
-                var bitmapImage = new BitmapImage();
-                bitmapImage.BeginInit();
-                bitmapImage.StreamSource = memory;
-                bitmapImage.CacheOption = BitmapCacheOption.OnLoad;
-                bitmapImage.EndInit();
-                bitmapImage.Freeze();
-
-                return bitmapImage;
-            }
-        }
     }
 }
