+++ EvoCarcassonne/EvoCarcassonne.csproj
@@ -77,11 +77,6 @@
     <Compile Include="Helpers\ObservableObject.cs" />
     <Compile Include="Helpers\RelayCommand.cs" />
     <Compile Include="Backend\Utils.cs" />
-    <Compile Include="Resources\Tiles.Designer.cs">
-      <AutoGen>True</AutoGen>
-      <DesignTime>True</DesignTime>
-      <DependentUpon>Tiles.resx</DependentUpon>
-    </Compile>
     <Compile Include="ViewModels\MainController.cs" />
     <Compile Include="Models\BoardTileModel.cs" />
     <Compile Include="ViewModels\ApplicationViewModel.cs" />
@@ -179,10 +174,6 @@
       <Generator>ResXFileCodeGenerator</Generator>
       <LastGenOutput>Resources.Designer.cs</LastGenOutput>
     </EmbeddedResource>
-    <EmbeddedResource Include="Resources\Tiles.resx">
-      <Generator>ResXFileCodeGenerator</Generator>
-      <LastGenOutput>Tiles.Designer.cs</LastGenOutput>
-    </EmbeddedResource>
     <None Include=".editorconfig" />
     <None Include="packages.config" />
     <None Include="Properties\Settings.settings">
@@ -201,6 +192,32 @@
   </ItemGroup>
   <ItemGroup>
     <Resource Include="Resources\backtile.png" />
+  </ItemGroup>
+  <ItemGroup>
+    <Resource Include="Resources\Tiles\S321014.png" />
+    <Resource Include="Resources\Tiles\T102020.png" />
+    <Resource Include="Resources\Tiles\T111113.png" />
+    <Resource Include="Resources\Tiles\T122021.png" />
+    <Resource Include="Resources\Tiles\T122123.png" />
+    <Resource Include="Resources\Tiles\T122221.png" />
+    <Resource Include="Resources\Tiles\T2001023.png" />
+    <Resource Include="Resources\Tiles\T202021.png" />
+    <Resource Include="Resources\Tiles\T220021.png" />
+    <Resource Include="Resources\Tiles\T221121.png" />
+    <Resource Include="Resources\Tiles\T222004.png" />
+    <Resource Include="Resources\Tiles\T2221213.png" />
+    <Resource Include="Resources\Tiles\T302024.png" />
+    <Resource Include="Resources\Tiles\T320020.png" />
+    <Resource Include="Resources\Tiles\T320114.png" />
+    <Resource Include="Resources\Tiles\T321104.png" />
+    <Resource Include="Resources\Tiles\T3211134.png" />
+    <Resource Include="Resources\Tiles\T321120.png" />
+    <Resource Include="Resources\Tiles\T322020.png" />
+    <Resource Include="Resources\Tiles\T400002.png" />
+    <Resource Include="Resources\Tiles\T401113.png" />
+    <Resource Include="Resources\Tiles\T520004.png" />
+    <Resource Include="Resources\Tiles\T810100.png" />
+    <Resource Include="Resources\Tiles\T900110.png" />
   </ItemGroup>
   <Import Project="$(MSBuildToolsPath)\Microsoft.CSharp.targets" />
 </Project>
