+++ EvoCarcassonne/Views/Modules/BottomControlModule.xaml
@@ -36,19 +36,16 @@
                 Width="24" />
             </Button>
 
-            <materialDesign:Card Grid.Column="2" Margin="10" Padding="-0.25">
+            <materialDesign:Card Grid.Column="2" Margin="10" Padding="-0.25" >
                 <Grid>
                     <Image Panel.ZIndex="1" Source="{Binding CurrentBoardTile.Image}">
                         <Image.LayoutTransform>
                             <TransformGroup>
-                                <ScaleTransform/>
-                                <SkewTransform/>
                                 <RotateTransform Angle="{Binding CurrentBoardTile.Angle}"/>
-                                <TranslateTransform/>
                             </TransformGroup>
                         </Image.LayoutTransform>
                     </Image>
-                    <Image Panel.ZIndex="-1" Source="/EvoCarcassonne;component/Resources/backtile.png"></Image>
+                    <Image Panel.ZIndex="-1" Source="/EvoCarcassonne;component/Resources/backtile.png" />
                 </Grid>
             </materialDesign:Card>
